fix: fix some bug with cohort-rights and pass biology research value in float
* release: release version 2.29.2

* fix: fix deidentied use case for date range into patient and encounter

* fix: fix some bug with cohort-rights and pass biology research value in float

+++ src/components/CreationCohort/DiagramView/components/LogicalOperator/components/CriteriaRightPanel/BiologyForm/components/Form/BiologyForm.tsx
@@ -248,7 +248,7 @@ const BiologyForm: React.FC<BiologyFormProps> = (props) => {
                 id="criteria-value"
                 variant="outlined"
                 value={currentState.valueMin}
-                onChange={(e) => onChangeValue('valueMin', parseInt(e.target.value))}
+                onChange={(e) => onChangeValue('valueMin', parseFloat(e.target.value))}
                 placeholder={currentState.valueComparator === Comparators.BETWEEN ? 'Valeur minimale' : '0'}
                 disabled={!allowSearchByValue}
               />
@@ -262,7 +262,7 @@ const BiologyForm: React.FC<BiologyFormProps> = (props) => {
                   id="criteria-value"
                   variant="outlined"
                   value={currentState.valueMax}
-                  onChange={(e) => onChangeValue('valueMax', parseInt(e.target.value))}
+                  onChange={(e) => onChangeValue('valueMax', parseFloat(e.target.value))}
                   placeholder="Valeur maximale"
                   disabled={!allowSearchByValue}
                 />
